refactor: themeConfig 모듈화
- themeConfig를 별도의 파일로 분리하여 모듈화
- Header, Footer 및 MainPage 컴포넌트를 각각의 폴더로 분리
